core: move extension loading logic from ServiceExtensionContext to ExtensionLoader
