+++ core/boot/src/main/java/org/eclipse/dataspaceconnector/boot/system/runtime/BaseRuntime.java
@@ -1,10 +1,28 @@
+/*
+ *  Copyright (c) 2020 - 2022 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Microsoft Corporation - initial API and implementation
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *
+ */
+
 package org.eclipse.dataspaceconnector.boot.system.runtime;
 
 
 import org.eclipse.dataspaceconnector.boot.monitor.MonitorProvider;
 import org.eclipse.dataspaceconnector.boot.system.DefaultServiceExtensionContext;
 import org.eclipse.dataspaceconnector.boot.system.ExtensionLoader;
+import org.eclipse.dataspaceconnector.boot.system.ServiceLocator;
+import org.eclipse.dataspaceconnector.boot.system.ServiceLocatorImpl;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
+import org.eclipse.dataspaceconnector.spi.system.ConfigurationExtension;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
 import org.eclipse.dataspaceconnector.spi.system.health.HealthCheckResult;
@@ -15,6 +33,7 @@ import org.eclipse.dataspaceconnector.spi.types.TypeManager;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
+import java.util.concurrent.ExecutorService;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
 
@@ -31,7 +50,7 @@ import static org.eclipse.dataspaceconnector.boot.system.ExtensionLoader.loadTel
  *     <li>{@link BaseRuntime#createContext(TypeManager, Monitor, Telemetry)}: creates a new {@link DefaultServiceExtensionContext} and invokes its {@link DefaultServiceExtensionContext#initialize()} method</li>
  *     <li>{@link BaseRuntime#initializeVault(ServiceExtensionContext)}: initializes the {@link org.eclipse.dataspaceconnector.spi.security.Vault} by
  *          calling {@link ExtensionLoader#loadVault(ServiceExtensionContext)} </li>
- *     <li>{@link BaseRuntime#createExtensions(ServiceExtensionContext)}: creates a list of {@code ServiceExtension} objects. By default, these are created through {@link ServiceExtensionContext#loadServiceExtensions()}</li>
+ *     <li>{@link BaseRuntime#createExtensions()}: creates a list of {@code ServiceExtension} objects. By default, these are created through {@link ServiceExtensionContext#loadServiceExtensions()}</li>
  *     <li>{@link BaseRuntime#bootExtensions(ServiceExtensionContext, List)}: initializes the service extensions by putting them through their lifecycle.
  *     By default this calls {@link ExtensionLoader#bootServiceExtensions(List, ServiceExtensionContext)} </li>
  *     <li>{@link BaseRuntime#onError(Exception)}: receives any Exception that was raised during initialization</li>
@@ -40,11 +59,22 @@ import static org.eclipse.dataspaceconnector.boot.system.ExtensionLoader.loadTel
 public class BaseRuntime {
 
     private final AtomicReference<HealthCheckResult> startupStatus = new AtomicReference<>(HealthCheckResult.failed("Startup not complete"));
+    private final ExtensionLoader extensionLoader;
+    protected final ServiceLocator serviceLocator;
     private Monitor monitor;
 
     public static void main(String[] args) {
         BaseRuntime runtime = new BaseRuntime();
         runtime.boot();
+    }
+
+    public BaseRuntime() {
+        this(new ServiceLocatorImpl());
+    }
+
+    protected BaseRuntime(ServiceLocator serviceLocator) {
+        this.extensionLoader = new ExtensionLoader(serviceLocator);
+        this.serviceLocator = serviceLocator;
     }
 
     protected Monitor getMonitor() {
@@ -60,7 +90,7 @@ public class BaseRuntime {
         var name = getRuntimeName(context);
         try {
             initializeVault(context);
-            List<InjectionContainer<ServiceExtension>> serviceExtensions = createExtensions(context);
+            List<InjectionContainer<ServiceExtension>> serviceExtensions = createExtensions();
             var seList = serviceExtensions.stream().map(InjectionContainer::getInjectionTarget).collect(Collectors.toList());
             getRuntime().addShutdownHook(new Thread(() -> shutdown(seList, monitor)));
             bootExtensions(context, serviceExtensions);
@@ -76,7 +106,6 @@ public class BaseRuntime {
         }
 
         monitor.info(format("%s ready", name));
-
     }
 
     @NotNull
@@ -132,11 +161,10 @@ public class BaseRuntime {
      * Create a list of {@link ServiceExtension}s. By default this is done using the ServiceLoader mechanism. Override if
      * e.g. a custom DI mechanism should be used.
      *
-     * @param context A context to which all the service extensions should be registered
      * @return a list of {@code ServiceExtension}s
      */
-    protected List<InjectionContainer<ServiceExtension>> createExtensions(ServiceExtensionContext context) {
-        return context.loadServiceExtensions();
+    protected List<InjectionContainer<ServiceExtension>> createExtensions() {
+        return extensionLoader.loadServiceExtensions();
     }
 
     /**
@@ -149,7 +177,11 @@ public class BaseRuntime {
      */
     @NotNull
     protected ServiceExtensionContext createContext(TypeManager typeManager, Monitor monitor, Telemetry telemetry) {
-        return new DefaultServiceExtensionContext(typeManager, monitor, telemetry);
+        return new DefaultServiceExtensionContext(typeManager, monitor, telemetry, loadConfigurationExtensions());
+    }
+
+    protected List<ConfigurationExtension> loadConfigurationExtensions() {
+        return extensionLoader.loadExtensions(ConfigurationExtension.class, false);
     }
 
     /**
@@ -180,7 +212,7 @@ public class BaseRuntime {
      * @param context An {@code ServiceExtensionContext} to resolve the {@code Vault} from.
      */
     protected void initializeVault(ServiceExtensionContext context) {
-        ExtensionLoader.loadVault(context);
+        ExtensionLoader.loadVault(context, extensionLoader);
     }
 
     /**
