+++ core/boot/src/test/java/org/eclipse/dataspaceconnector/boot/system/ExtensionLoaderTest.java
@@ -9,6 +9,7 @@
  *
  *  Contributors:
  *       Microsoft Corporation - initial API and implementation
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
  *
  */
 
@@ -16,15 +17,25 @@ package org.eclipse.dataspaceconnector.boot.system;
 
 import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.opentelemetry.api.OpenTelemetry;
+import org.eclipse.dataspaceconnector.boot.util.CyclicDependencyException;
+import org.eclipse.dataspaceconnector.spi.EdcException;
 import org.eclipse.dataspaceconnector.spi.monitor.ConsoleMonitor;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
 import org.eclipse.dataspaceconnector.spi.monitor.MultiplexingMonitor;
 import org.eclipse.dataspaceconnector.spi.security.CertificateResolver;
 import org.eclipse.dataspaceconnector.spi.security.PrivateKeyResolver;
 import org.eclipse.dataspaceconnector.spi.security.Vault;
+import org.eclipse.dataspaceconnector.spi.system.BaseExtension;
+import org.eclipse.dataspaceconnector.spi.system.Inject;
 import org.eclipse.dataspaceconnector.spi.system.MonitorExtension;
+import org.eclipse.dataspaceconnector.spi.system.Provides;
+import org.eclipse.dataspaceconnector.spi.system.Requires;
+import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
 import org.eclipse.dataspaceconnector.spi.system.VaultExtension;
+import org.eclipse.dataspaceconnector.spi.system.injection.EdcInjectionException;
+import org.eclipse.dataspaceconnector.spi.system.injection.InjectionContainer;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
@@ -32,10 +43,12 @@ import java.util.List;
 
 import static java.util.Collections.emptyList;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyBoolean;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.ArgumentMatchers.isA;
 import static org.mockito.Mockito.atLeastOnce;
@@ -45,6 +58,10 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 class ExtensionLoaderTest {
+
+    private final ServiceLocator serviceLocator = mock(ServiceLocator.class);
+    private final ServiceExtension coreExtension = new TestCoreExtension();
+    private final ExtensionLoader loader = new ExtensionLoader(serviceLocator);
 
     @BeforeAll
     public static void setup() {
@@ -112,15 +129,15 @@ class ExtensionLoaderTest {
         DefaultServiceExtensionContext contextMock = mock(DefaultServiceExtensionContext.class);
 
         when(contextMock.getMonitor()).thenReturn(mock(Monitor.class));
-        when(contextMock.loadSingletonExtension(VaultExtension.class, false)).thenReturn(null);
+        when(serviceLocator.loadSingletonImplementor(VaultExtension.class, false)).thenReturn(null);
 
-        ExtensionLoader.loadVault(contextMock);
+        ExtensionLoader.loadVault(contextMock, loader);
 
         verify(contextMock).registerService(eq(Vault.class), isA(Vault.class));
         verify(contextMock).registerService(eq(PrivateKeyResolver.class), any());
         verify(contextMock).registerService(eq(CertificateResolver.class), any());
         verify(contextMock, atLeastOnce()).getMonitor();
-        verify(contextMock).loadSingletonExtension(VaultExtension.class, false);
+        verify(serviceLocator).loadSingletonImplementor(VaultExtension.class, false);
     }
 
     @Test
@@ -130,7 +147,7 @@ class ExtensionLoaderTest {
         PrivateKeyResolver resolverMock = mock(PrivateKeyResolver.class);
         CertificateResolver certResolverMock = mock(CertificateResolver.class);
         when(contextMock.getMonitor()).thenReturn(mock(Monitor.class));
-        when(contextMock.loadSingletonExtension(VaultExtension.class, false)).thenReturn(new VaultExtension() {
+        when(serviceLocator.loadSingletonImplementor(VaultExtension.class, false)).thenReturn(new VaultExtension() {
 
             @Override
             public Vault getVault() {
@@ -148,10 +165,232 @@ class ExtensionLoaderTest {
             }
         });
 
-        ExtensionLoader.loadVault(contextMock);
+        ExtensionLoader.loadVault(contextMock, loader);
 
         verify(contextMock, times(1)).registerService(Vault.class, vaultMock);
         verify(contextMock, atLeastOnce()).getMonitor();
-        verify(contextMock).loadSingletonExtension(VaultExtension.class, false);
+        verify(serviceLocator).loadSingletonImplementor(VaultExtension.class, false);
+    }
+
+    @Test
+    @DisplayName("No dependencies between service extensions")
+    void loadServiceExtensions_noDependencies() {
+
+        var service1 = new ServiceExtension() {
+        };
+        when(serviceLocator.loadImplementors(eq(ServiceExtension.class), anyBoolean())).thenReturn(mutableListOf(service1, coreExtension));
+
+        var list = loader.loadServiceExtensions();
+
+        assertThat(list).hasSize(2);
+        assertThat(list).extracting(InjectionContainer::getInjectionTarget).contains(service1);
+        verify(serviceLocator).loadImplementors(eq(ServiceExtension.class), anyBoolean());
+    }
+
+    @Test
+    @DisplayName("Locating two service extensions for the same service class ")
+    void loadServiceExtensions_whenMultipleServices() {
+        var service1 = new ServiceExtension() {
+        };
+        var service2 = new ServiceExtension() {
+        };
+
+        when(serviceLocator.loadImplementors(eq(ServiceExtension.class), anyBoolean())).thenReturn(mutableListOf(service1, service2, coreExtension));
+
+        var list = loader.loadServiceExtensions();
+        assertThat(list).hasSize(3);
+        assertThat(list).extracting(InjectionContainer::getInjectionTarget).containsExactlyInAnyOrder(service1, service2, coreExtension);
+        verify(serviceLocator).loadImplementors(eq(ServiceExtension.class), anyBoolean());
+    }
+
+    @Test
+    @DisplayName("A DEFAULT service extension depends on a PRIMORDIAL one")
+    void loadServiceExtensions_withBackwardsDependency() {
+        var depending = new DependingExtension();
+        var someExtension = new SomeExtension();
+        var providing = new ProvidingExtension();
+
+        when(serviceLocator.loadImplementors(eq(ServiceExtension.class), anyBoolean())).thenReturn(mutableListOf(providing, depending, someExtension, coreExtension));
+
+        var services = loader.loadServiceExtensions();
+        assertThat(services).extracting(InjectionContainer::getInjectionTarget).containsExactly(coreExtension, providing, depending, someExtension);
+        verify(serviceLocator).loadImplementors(eq(ServiceExtension.class), anyBoolean());
+    }
+
+
+    @Test
+    @DisplayName("A service extension has a dependency on another one of the same loading stage")
+    void loadServiceExtensions_withEqualDependency() {
+        var depending = new DependingExtension() {
+        };
+        var coreService = new SomeExtension() {
+        };
+
+        var thirdService = new ServiceExtension() {
+        };
+
+        when(serviceLocator.loadImplementors(eq(ServiceExtension.class), anyBoolean())).thenReturn(mutableListOf(depending, thirdService, coreService, coreExtension));
+
+        var services = loader.loadServiceExtensions();
+        assertThat(services).extracting(InjectionContainer::getInjectionTarget).containsExactlyInAnyOrder(coreService, depending, thirdService, coreExtension);
+        verify(serviceLocator).loadImplementors(eq(ServiceExtension.class), anyBoolean());
+    }
+
+    @Test
+    @DisplayName("Two service extensions have a circular dependency")
+    void loadServiceExtensions_withCircularDependency() {
+        var s1 = new TestProvidingExtension2();
+        var s2 = new TestProvidingExtension();
+
+        when(serviceLocator.loadImplementors(eq(ServiceExtension.class), anyBoolean())).thenReturn(mutableListOf(s1, s2, coreExtension));
+
+        assertThatThrownBy(() -> loader.loadServiceExtensions()).isInstanceOf(CyclicDependencyException.class);
+        verify(serviceLocator).loadImplementors(eq(ServiceExtension.class), anyBoolean());
+    }
+
+    @Test
+    @DisplayName("A service extension has an unsatisfied dependency")
+    void loadServiceExtensions_dependencyNotSatisfied() {
+        var depending = new DependingExtension();
+        var someExtension = new SomeExtension();
+
+        when(serviceLocator.loadImplementors(eq(ServiceExtension.class), anyBoolean())).thenReturn(mutableListOf(depending, someExtension, coreExtension));
+
+        assertThatThrownBy(() -> loader.loadServiceExtensions()).isInstanceOf(EdcException.class).hasMessageContaining("The following injected fields were not provided:\nField \"someService\" of type ");
+        verify(serviceLocator).loadImplementors(eq(ServiceExtension.class), anyBoolean());
+    }
+
+    @Test
+    @DisplayName("Services extensions are sorted by dependency order")
+    void loadServiceExtensions_dependenciesAreSorted() {
+        var depending = new DependingExtension();
+        var providingExtension = new ProvidingExtension();
+
+
+        when(serviceLocator.loadImplementors(eq(ServiceExtension.class), anyBoolean())).thenReturn(mutableListOf(depending, providingExtension, coreExtension));
+
+        var services = loader.loadServiceExtensions();
+        assertThat(services).extracting(InjectionContainer::getInjectionTarget).containsExactly(coreExtension, providingExtension, depending);
+        verify(serviceLocator).loadImplementors(eq(ServiceExtension.class), anyBoolean());
+    }
+
+    @Test
+    @DisplayName("Should throw exception when no core dependency found")
+    void loadServiceExtensions_noCoreDependencyShouldThrowException() {
+        var depending = new DependingExtension();
+        var coreService = new SomeExtension();
+        when(serviceLocator.loadImplementors(eq(ServiceExtension.class), anyBoolean())).thenReturn(mutableListOf(depending, coreService));
+
+        assertThatThrownBy(() -> loader.loadServiceExtensions()).isInstanceOf(EdcException.class);
+    }
+
+    @Test
+    @DisplayName("Requires annotation influences ordering")
+    void loadServiceExtensions_withAnnotation() {
+        var depending = new DependingExtension();
+        var providingExtension = new ProvidingExtension();
+        var annotatedExtension = new AnnotatedExtension();
+        when(serviceLocator.loadImplementors(eq(ServiceExtension.class), anyBoolean())).thenReturn(mutableListOf(depending, annotatedExtension, providingExtension, coreExtension));
+
+        var services = loader.loadServiceExtensions();
+
+        assertThat(services).extracting(InjectionContainer::getInjectionTarget).containsExactly(coreExtension, providingExtension, depending, annotatedExtension);
+        verify(serviceLocator).loadImplementors(eq(ServiceExtension.class), anyBoolean());
+    }
+
+    @Test
+    @DisplayName("Requires annotation not satisfied")
+    void loadServiceExtensions_withAnnotation_notSatisfied() {
+        var annotatedExtension = new AnnotatedExtension();
+
+        when(serviceLocator.loadImplementors(eq(ServiceExtension.class), anyBoolean())).thenReturn(mutableListOf(annotatedExtension, coreExtension));
+
+        assertThatThrownBy(() -> loader.loadServiceExtensions()).isNotInstanceOf(EdcInjectionException.class).isInstanceOf(EdcException.class);
+        verify(serviceLocator).loadImplementors(eq(ServiceExtension.class), anyBoolean());
+    }
+
+    @Test
+    @DisplayName("Mixed requirement features work")
+    void loadServiceExtensions_withMixedInjectAndAnnotation() {
+        var providingExtension = new ProvidingExtension(); // provides SomeObject
+        var anotherProvidingExt = new AnotherProvidingExtension(); //provides AnotherObject
+        var mixedAnnotation = new MixedAnnotation();
+
+        when(serviceLocator.loadImplementors(eq(ServiceExtension.class), anyBoolean())).thenReturn(mutableListOf(mixedAnnotation, providingExtension, coreExtension, anotherProvidingExt));
+
+        var services = loader.loadServiceExtensions();
+        assertThat(services).extracting(InjectionContainer::getInjectionTarget).containsExactly(coreExtension, providingExtension, anotherProvidingExt, mixedAnnotation);
+        verify(serviceLocator).loadImplementors(eq(ServiceExtension.class), anyBoolean());
+    }
+
+    @Test
+    @DisplayName("Mixed requirement features introducing circular dependency")
+    void loadServiceExtensions_withMixedInjectAndAnnotation_withCircDependency() {
+        var s1 = new TestProvidingExtension3();
+        var s2 = new TestProvidingExtension();
+
+        when(serviceLocator.loadImplementors(eq(ServiceExtension.class), anyBoolean())).thenReturn(mutableListOf(s1, s2, coreExtension));
+
+        assertThatThrownBy(() -> loader.loadServiceExtensions()).isInstanceOf(CyclicDependencyException.class);
+        verify(serviceLocator).loadImplementors(eq(ServiceExtension.class), anyBoolean());
+    }
+
+    @SafeVarargs
+    private <T> List<T> mutableListOf(T... elements) {
+        return new ArrayList<>(List.of(elements));
+    }
+
+    private static class DependingExtension implements ServiceExtension {
+        @Inject
+        private SomeObject someService;
+    }
+
+    private static class SomeExtension implements ServiceExtension {
+    }
+
+    @Provides({ SomeObject.class })
+    private static class ProvidingExtension implements ServiceExtension {
+    }
+
+    @Provides(AnotherObject.class)
+    private static class AnotherProvidingExtension implements ServiceExtension {
+    }
+
+    @Provides({ SomeObject.class })
+    private static class TestProvidingExtension implements ServiceExtension {
+        @Inject
+        AnotherObject obj;
+    }
+
+    @Provides({ AnotherObject.class })
+    private static class TestProvidingExtension2 implements ServiceExtension {
+        @Inject
+        SomeObject obj;
+    }
+
+    @Provides({ AnotherObject.class })
+    @Requires({ SomeObject.class })
+    private static class TestProvidingExtension3 implements ServiceExtension {
+    }
+
+    @Requires(SomeObject.class)
+    private static class AnnotatedExtension implements ServiceExtension {
+    }
+
+    @Requires(SomeObject.class)
+    private static class MixedAnnotation implements ServiceExtension {
+        @Inject
+        private AnotherObject obj;
+    }
+
+    private static class SomeObject {
+    }
+
+    private static class AnotherObject {
+    }
+
+    @BaseExtension
+    private static class TestCoreExtension implements ServiceExtension {
+
     }
 }
