+++ samples/other/custom-runtime/src/main/java/org/eclipse/dataspaceconnector/demo/runtime/CustomRuntime.java
@@ -3,6 +3,7 @@ package org.eclipse.dataspaceconnector.demo.runtime;
 import org.eclipse.dataspaceconnector.boot.system.DefaultServiceExtensionContext;
 import org.eclipse.dataspaceconnector.boot.system.runtime.BaseRuntime;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
+import org.eclipse.dataspaceconnector.spi.system.ConfigurationExtension;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
 import org.eclipse.dataspaceconnector.spi.telemetry.Telemetry;
@@ -10,8 +11,6 @@ import org.eclipse.dataspaceconnector.spi.types.TypeManager;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
-
-import static org.eclipse.dataspaceconnector.boot.system.ExtensionLoader.loadTelemetry;
 
 public class CustomRuntime extends BaseRuntime {
 
@@ -31,7 +30,7 @@ public class CustomRuntime extends BaseRuntime {
     @Override
     protected @NotNull ServiceExtensionContext createContext(TypeManager typeManager, Monitor monitor, Telemetry telemetry) {
         //override the default service extension context with a super customized one
-        return new SuperCustomExtensionContext(typeManager, monitor, telemetry);
+        return new SuperCustomExtensionContext(typeManager, monitor, telemetry, loadConfigurationExtensions());
     }
 
     @Override
@@ -43,8 +42,8 @@ public class CustomRuntime extends BaseRuntime {
     }
 
     private static class SuperCustomExtensionContext extends DefaultServiceExtensionContext {
-        public SuperCustomExtensionContext(TypeManager typeManager, Monitor monitor, Telemetry telemetry) {
-            super(typeManager, monitor, telemetry);
+        public SuperCustomExtensionContext(TypeManager typeManager, Monitor monitor, Telemetry telemetry, List<ConfigurationExtension> configurationExtensions) {
+            super(typeManager, monitor, telemetry, configurationExtensions);
         }
     }
 }
